GrowableRoot
The root can now grow. It's all on the shader GrowRoots. Just need to integrate the triggering of it into the Manager.  I added the new scene MainGame-(HenryEDIT) - It also has some microbes in it and more fungi. Also if you open the New-MAIN ROOT model you will see there some fungi models to be activated and turned on when the the player has connected the plant to the web